pyproject.toml
[project]
name = "autonews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "feedparser>=6.0.12",
    "prefect>=3.4.25",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-ai-slim[openai,openrouter]>=0.8.1",
    "sqlmodel>=0.0.34",
    "trafilatura>=2.0.0",
]
